test: fix tests and rename files
